Feature/dotnet publish env var (#163)
* Added publishEnvironmentVariables to dotnet publish step

* Updated readme

* Enhance DotNet build with environment variables

Added environment variable support for DotNet build.

* Update build_jobs.yml

* Add publishEnvironmentVariables parameter to build stages

* Update README to include publishEnvironmentVariables

Added 'publishEnvironmentVariables' flag to README.

* Update README.md

* Refactor publishEnvironmentVariables format in README

Updated the format of publishEnvironmentVariables in README.

* Update README to change publishEnvironmentVariables type to object

Modified the publishEnvironmentVariables entry in both dotnet and dotnetweb README files to reflect its new type as an object, allowing for a dictionary of environment variables to be published during deployment.

* Clarify publishEnvironmentVariables description in README files

Updated the description of the publishEnvironmentVariables entry in both dotnet and dotnetweb README files to specify that it is used to pass environment variables to the build and publish tasks, respectively.

* Fixed buildEnvironmentVariables

* buildEnvironmentVariables